+++ taichi/backends/vulkan/aot_module_builder_impl.h
@@ -5,7 +5,7 @@
 
 #include "taichi/aot/module_builder.h"
 #include "taichi/backends/vulkan/aot_utils.h"
-#include "taichi/backends/vulkan/runtime.h"
+#include "taichi/runtime/vulkan/runtime.h"
 #include "taichi/codegen/spirv/snode_struct_compiler.h"
 #include "taichi/codegen/spirv/kernel_utils.h"
